Refactor workshop final into package
